Update changelog
In [PR 2088][] we introduced "user capabilities" to overcome the
limitations of the complex rules in how certain key user attributes are
assigned. This code is called every time a user is saved:

```
private def apply_roles_based_on_team
  assign_attributes(
    assign_to_project: allowing_override_for(:assign_to_project) { is_regional_caseworker? || is_regional_delivery_officer? },
    manage_user_accounts: apply_service_support_role?,
    manage_conversion_urns: apply_service_support_role?,
    manage_local_authorities: apply_service_support_role?,
    add_new_project: allowing_override_for(:add_new_project) { is_regional_delivery_officer? },
    manage_team: allowing_override_for(:manage_team) { apply_team_lead_role? }
  )
end
```

Our user capabilities allow for the following to be overridden using
user capabilities:

- `add_new_project`
- `assign_to_project`
- `manage_team`

In [PR 2138][] we added a further capability (`devops`) to allow users
outside of the "service support" team to access the page which shows
the current "dual running" config on the Azure FrontDoor and its effect
on the service's routes.

In this change we provide a UI for members of the service support team
to view and manage these capabilities.

[PR 2088]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/2088

[PR 2138]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/2138/commits/f6f50e18965cbcc3f611b768858fa50c7f023b7f