Add not-null constraints to the notes table
We're seeing a lot of invalid notes being created in the dev and
test dbs by the new .NET version of the service which is under
construction.

In the Ruby version, the `User` and `Project` associations are enforced,
as is the presence of the `#body`.

In this migration we add these constraints to the `notes` table in the db
itself.