Respect TRN validation when editing
The `FormAMultiAcademyTrustNameValidator` is used in the
"CreateProject" form object. Now that we are allowing the TRN
to  be changed we use that validator here in the "EditProject"
version of the form object to ensure that all projects which
share a `new_trust_reference_number` have the same
`new_trust_name`.

In the future we ought to refactor this out to a separate
`ActiveRecord` class and db table, e.g. `NewTrustReference`
with (`:name` and `:number`).