Fix flaky test link to edit sharepoint link
We're going to be introducing a new link here to edit
the TRN. In our feature tests we'll be addressing it
using a class, not via this "position in list" technique
which has no meaning and is brittle.

See Jim Weirich's talk [Connascence Examined][] from ETE 2012 or
[Connascence as a Software Design Metric][] 2011 by Gregory Brown

The point is that referring to something using a name is
better than referring to it by its position in a list.

[Connasence Examined]:
https://www.youtube.com/watch?v=HQXVKHoUQxY

[Connascence as a Software Design Metric]:
https://practicingruby.com/articles/connascence