Ensure that a transfer Form a Mat respects validation rules
As we did earlier with the `Conversion::EditProjectForm` we now apply
the `FormAMultiAcademyTrustNameValidator` to the `Transfer::EditProjectForm`
to ensure that all projects which share a `new_trust_reference_number`
have the same `new_trust_name`.

In the future we ought to refactor this out to a separate
`ActiveRecord` class and db table, e.g. `NewTrustReference`
with (`:name` and `:number`).